Fix parameter type in SQLStmt::bind
This should have been an integer from the beginning, not sure why I
missed that. Apparently this did work (?) with PHP 7, but now with PHP 8
it actually converts the argument to string and passing a string to
SQLite3Stmt::bindValue identifies it as named parameter, which is not
what we want.